See desc for all changes
FVN.dll: updated AddTroop and Propogate to actually pass in floats, not ints.
Peasant prefab: Enabled Atk/Mv scripts, added bools to check whether they should run.
ALL SCRIPTS: Added namespaces to help with code structure and refactoring for testing
PlayerAttack: changed camera reference to main scene cam, added in-script references to spotlight/gameloop, restructured Update loop checks, added fudge roll to Attack(), moved all UI setting to UI Controller, moved all spotlight setting to gameloop
PlayerMagic: added bool to check whether explosions were being placed
PlayerMovement: moved update body to separate function, added boolean for Update()
Character: started implementing state machine to simplify state switching
CharacterFeatures: removed monobehaviour, easier to test now, moved color changing to PlayerSpotlight
GameLoop: Moved to new folder, presumably moved stuff around but I can't tell
InputManager: Returns current input, meaning we can set the result in tests
BattleUIControl: Presumably there were changes, but moving to a new folder wiped them out
PlayerSpotlight: As above
Client: Minor adjustments, disabled warnings in Log, restructuring flow of info
TestGamePlay: put common test start in single method which gets auto called before each test
Code Coverage: I ran the code coverage tool, which deleted everything and set it back again.
----SERVER----
Changed DB interactions to actually use float for stored positions.

+++ FearValleyNetworking/Net_AddTroop.cs
@@ -21,8 +21,8 @@
         /// <summary>Armor carried by troop.</summary>
         public string ArmorType { set; get; }
         /// <summary>X position where they were placed, relative to thier perspective.</summary>
-        public int XPosRelative { set; get; }
+        public float XPosRelative { set; get; }
         /// <summary>Z position where they were placed, relative to their perspective.</summary>
-        public int ZPosRelative { set; get; }
+        public float ZPosRelative { set; get; }
     }
 }
